% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lbfs.R
\name{calc_lbf}
\alias{calc_lbf}
\title{Calculate Laplace Bayes factors.}
\usage{
calc_lbf(beta_hat, v, lambda)
}
\arguments{
\item{beta_hat}{The effect size estimate}

\item{v}{The variance of the effect size estimate.}

\item{lambda}{The rate parameter of the Laplace prior.}
}
\description{
Calculate Laplace Bayes factors.
}
\examples{
calc_lbf(beta_hat = 0.1, v = 0.6, lambda = 5)
}
