% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaries_multi_data.R
\name{calc_summary}
\alias{calc_summary}
\title{Calculate Wakefield and Laplace Bayes factors.}
\usage{
calc_summary(
  log_reg_results_array,
  lambda,
  num_snps,
  num_datasets,
  calc_name,
  w
)
}
\arguments{
\item{log_reg_results_array}{a 3 dimensional array of dimension num_snps by
2 by num_datasets. Column 1 contains the estimates of the effect sizes for
each SNP, column 2 contains the variances of the effect size estimates for
each SNP in each dataset.}

\item{lambda}{The rate parameter of the Laplace prior.}

\item{num_snps}{The number of SNPs to be fine mapped.}

\item{num_datasets}{The number of datasets being supplied (designed for use
with simulated data from, for example, Hapgen).}

\item{calc_name}{A character string representing the statistic to be used.
Possibilities are lbf and wbf for Laplace and Wakefield Bayes factor; mean,
median and mode for posterior mean, median and mode; ci and hdi for 95
percent posterior credible or high density interval.}

\item{w}{The variance of the Gaussian effect size prior}
}
\description{
Calculate Wakefield and Laplace Bayes factors.
}
