% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wbfs.R
\name{calc_wbf}
\alias{calc_wbf}
\title{Calculate Wakefield Bayes factors.}
\usage{
calc_wbf(beta_hat, v, w)
}
\arguments{
\item{beta_hat}{The effect size estimate}

\item{v}{The variance of the effect size estimate.}

\item{w}{The variance of the Gaussian effect size prior}
}
\description{
Calculate Wakefield Bayes factors.
}
\examples{
calc_wbf(beta_hat = 0.1, v = 0.6, w = 0.04)
}
