% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_credible_sets.R
\name{find_cred_set}
\alias{find_cred_set}
\alias{find_ppc}
\title{Find credible sets}
\usage{
find_cred_set(ppc, causal_snp, thresh)

find_ppc(bf)
}
\arguments{
\item{ppc}{A vector of posterior probabilities of causality
(defined as Bayes factor/sum of Bayes factors).}

\item{causal_snp}{The causal SNP number where SNPs are numbered from one to
the number of SNPs.}

\item{thresh}{The cumulative probability threshold used to determine the
credible set.}

\item{bf}{vector of Bayes factors.}
}
\value{
find_ppc returns the vector of posterior probabilities of causality
for each SNP; find_cred_set returns a length 2 vector where the first element
is the size of the credible set and the second element is T or F depending on
whether the credible set contains the causal SNP.
}
\description{
Calculate posterior probabilities of causality and form credible sets
assuming there is a single causal SNP.
}
\section{Functions}{
\itemize{
\item \code{find_ppc}: Calculate posterior probabilities of causality
}}

\examples{
find_cred_set(ppc = c(0.10, 0.20, 0.35, 0.40), causal_snp = 3, thresh = 0.9)
}
