% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noteworthy.R
\name{find_noteworthy}
\alias{find_noteworthy}
\title{Determine which SNPS are noteworthy.}
\usage{
find_noteworthy(bf, pi_0, r)
}
\arguments{
\item{bf}{A vector of Bayes factors.}

\item{pi_0}{The prior probability of no association.}

\item{r}{The ratio of the costs of false non-discovery to that of false discovery.}
}
\description{
Determine which SNPS are noteworthy.
}
