% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_logreg_multidata.R
\name{run_logistic_regression}
\alias{run_logistic_regression}
\title{Perform univariate logistic regression on case-control genotype data.}
\usage{
run_logistic_regression(
  geno_array,
  num_cases,
  num_controls,
  num_snps,
  num_datasets
)
}
\arguments{
\item{geno_array}{A (\code{num_cases} + \code{num_controls}) by \code{num_snps} by
\code{num_datasets} array of genotypes.}

\item{num_cases}{The number of cases in \code{geno_array}.}

\item{num_controls}{The number of controls in \code{geno_array}.}

\item{num_snps}{The number of SNPs you are expecting \code{geno_array} to contain.}

\item{num_datasets}{The number of datasets you are expecting \code{geno_array} to contain.}
}
\value{
A \code{num_snps} by three by \code{num_datasets} array. The three
  values of the second element of the array are the mle, variance of the mle
  and the p-value from a univariate logistic regression analysis (with null
  hypothesis that the effect size is zero).
}
\description{
Perform univariate logistic regression on case-control genotype data.
}
